Move more test related code into testutil
