+++ vault/resource_ad_secret_backend_test.go
@@ -11,12 +11,11 @@ import (
 	"github.com/hashicorp/vault/api"
 
 	"github.com/hashicorp/terraform-provider-vault/testutil"
-	"github.com/hashicorp/terraform-provider-vault/util"
 )
 
 func TestADSecretBackend(t *testing.T) {
 	backend := acctest.RandomWithPrefix("tf-test-ad")
-	bindDN, bindPass, url := util.GetTestADCreds(t)
+	bindDN, bindPass, url := testutil.GetTestADCreds(t)
 
 	resource.Test(t, resource.TestCase{
 		Providers:                 testProviders,
